Fooled around in userStorage
We added the functions of EditUser s.. this is still just a sketch... Okay, today we got an example of Linus how to store a list of to do's in local storage. Itäs just a model fpor how to operate. I told him later we have created the following structure.. We have a list of users in which every user has several properties - amongst other an array full of to do's .. He said that it actualy quite smart! So we continue along this road. So today we learned that each time we change our to do list we need to update our to do list in local storarge.. So in our case this means: every time we have updated a to do -list of a user, we have updated an object in the the user list, so every time we need to update store our user list in the local storage ...